webMAN MOD 1.47.28 + video_rec plugin
- The standalone video_rec plugin now allows to select the audio/video format
- The web command /browse.ps3?<url> now is allowed in-game
- PSXISO with more than 32 tracks now load with 32 tracks instead of return error
- Removed unnecessary remap
- /remap.ps3 without parameters now reset all the remaps
- Renamed function set_apphome() to set_app_home()
- /play.ps3/app_home now starts the debug icon

+++ _Projects_/ps3netsrv/src/VIsoFile.cpp
@@ -1662,7 +1662,7 @@ ssize_t VIsoFile::read(void *buf, size_t nbyte)
 							return r;
 						}
 
-						if (this_r != to_read)
+						if (this_r != (int64_t)to_read)
 						{
 							fprintf(stderr, "VISO: read on file %s returned less data than expected (file modified?)\n", fileList->path);
 							return r;
